eina: use uintptr_t for Eina_Thread storage to be the same size for the different architectures.
This should not be an ABI regression on the platform where it work (Unix). It is on Windows,
but the API was unusable at all without this change, so this is a necessary breakage to
unbreak things.

@fix

+++ src/lib/eina/eina_thread.h
@@ -23,6 +23,8 @@
 #include "eina_types.h"
 #include "eina_error.h"
 
+#include <stdint.h>
+
 /**
  * @addtogroup Eina_Tools_Group Tools
  *
@@ -46,7 +48,7 @@
  * @typedef Eina_Thread
  * Type for a generic thread.
  */
-typedef unsigned long int Eina_Thread;
+typedef uintptr_t Eina_Thread;
 
 /**
  * @typedef Eina_Thread_Cb
